+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+- Fixed $refs with percent-encoded paths
 
 - Fixed #1636. multiselect obstructed editor format "table"
 - Fixed #1559 and #1621 field dependent on false should now display